Minor fixes in networking code

+++ src-api/core/olsr_packet_socket.h
@@ -94,7 +94,7 @@ struct olsr_packet_managed_config {
   char interface[IF_NAMESIZE];
   struct netaddr bindto_v4, multicast_v4;
   struct netaddr bindto_v6, multicast_v6;
-  uint16_t port, multicast_port;
+  int port, multicast_port;
   bool loop_multicast;
 };
